refactor: 아이콘 컴포넌트 스타일, 타입, 컴포넌트 분리
