+++ client/src/components/Icon/Icon.style.ts
@@ -1,0 +1,7 @@
+import styled from "styled-components";
+
+const Container = styled.span`
+  display: inline-block;
+`;
+
+export default Container;
