+++ client/src/components/Icon/Icon.type.ts
@@ -1,0 +1,6 @@
+interface IconProps {
+  type: "plus" | "kebab" | "trashcan" | "text";
+  size: string;
+}
+
+export default IconProps;
